requirements.txt
absl-py==2.1.0
aiodns==3.1.1
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==25.3.0
baron==0.10.1
beautifulsoup4==4.12.2
blessed==1.20.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
codefind==0.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
discord==2.3.2
discord.py==2.5.2
distlib==0.3.6
dj-database-url==1.3.0
dj-email-url==1.0.6
Django==3.2.18
django-cache-url==3.4.4
django-environ==0.10.0
django-environs==0.1.1
django-extensions==3.2.1
django-import-export==3.1.0
environs==9.5.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.9.0
Flask==2.2.3
Flask-Compress==1.13
flatbuffers==24.3.25
fonttools==4.47.0
frozenlist==1.3.3
future==0.18.3
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
idna==3.4
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jax==0.4.26
jedi==0.19.1
Jinja2==3.1.2
jinxed==1.2.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.1
kiwisolver==1.4.5
labelImg==1.8.6
libclang==18.1.1
line-bot-sdk==2.2.1
lxml==6.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
multidict==6.0.4
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.2
odfpy==1.4.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.1
opt-einsum==3.3.0
optree==0.11.0
orjson==3.9.14
outcome==1.3.0.post0
ovld==0.3.2
packaging==23.0
pandas==2.2.2
parso==0.8.3
Pillow==10.1.0
platformdirs==3.1.0
plotly==5.13.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
py-cord @ git+https://github.com/Pycord-Development/pycord@845f7c7d7250887d62bbf65d76468e5d45263773
pycares==4.4.0
pycparser==2.21
Pygments==2.17.2
pyhibp==4.2.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
redbaron==0.9.2
requests==2.28.2
rich==13.7.1
rply==0.7.8
scipy==1.13.0
selenium==4.30.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.3
tablib==3.3.0
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.4
traitlets==5.14.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2022.7
urllib3==1.26.15
uv==0.8.13
virtualenv==20.20.0
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
